refatora syntaxBuilder
